Move props out of privider module
