+++ provider/data/properties/stubdata/hangul_syllable_type_v1_marker.rs.data
@@ -16,7 +16,7 @@ macro_rules! __impl_hangul_syllable_type_v1_marker {
         #[clippy::msrv = "1.71.1"]
         impl $provider {
             #[doc(hidden)]
-            pub const SINGLETON_HANGUL_SYLLABLE_TYPE_V1_MARKER: &'static <icu::properties::provider::HangulSyllableTypeV1Marker as icu_provider::DynamicDataMarker>::DataStruct = &icu::properties::provider::PropertyCodePointMapV1::CodePointTrie(icu::collections::codepointtrie::CodePointTrie::from_parts(icu::collections::codepointtrie::CodePointTrieHeader { high_start: 55296u32, shifted12_high_start: 14u16, index3_null_offset: 2u16, data_null_offset: 0u32, null_value: 0u32, trie_type: icu::collections::codepointtrie::TrieType::Small }, unsafe { zerovec::ZeroVec::from_bytes_unchecked(b"\0\0@\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\xDA\0\xE3\0\xEF\0\x08\x01\0\0\x10\0 \x000\0@\0P\0`\0p\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\x80\0\x80\0\x80\0\x80\0\x80\0\x80\0\x90\0\x90\0\x90\0\x90\0\x98\0\xA0\0\xA0\0\xA0\0\xA0\0\xA0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\x80\0\xB0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xDD\0\x90\0\xED\0\xA0\0\xA0\0\xF8\0D\0L\0L\0L\0L\0L\0L\0L\0l\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x8C\0\x02\0\xAC\0\xB0\0\xAD\0\xB1\0\xAE\0\xB2\0\xAF\0\xAC\0\xB0\0\xAD\0\xB1\0\xAE\0\xB2\0\xAF\0\xBA\0") }, unsafe { zerovec::ZeroVec::from_bytes_unchecked(b"\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\0\0\0\x04\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x04\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\0\0\0\0\0\0\0\0\0\0\0\0\x02\x02\x02\x02\x02\x02\x02\0\0\0\0\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\0\0\0\0") }, icu::properties::provider::props::HangulSyllableType(0u8)));
+            pub const SINGLETON_HANGUL_SYLLABLE_TYPE_V1_MARKER: &'static <icu::properties::provider::HangulSyllableTypeV1Marker as icu_provider::DynamicDataMarker>::DataStruct = &icu::properties::provider::PropertyCodePointMapV1::CodePointTrie(icu::collections::codepointtrie::CodePointTrie::from_parts(icu::collections::codepointtrie::CodePointTrieHeader { high_start: 55296u32, shifted12_high_start: 14u16, index3_null_offset: 2u16, data_null_offset: 0u32, null_value: 0u32, trie_type: icu::collections::codepointtrie::TrieType::Small }, unsafe { zerovec::ZeroVec::from_bytes_unchecked(b"\0\0@\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\xDA\0\xE3\0\xEF\0\x08\x01\0\0\x10\0 \x000\0@\0P\0`\0p\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\x10\0 \x000\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\x80\0\x80\0\x80\0\x80\0\x80\0\x80\0\x90\0\x90\0\x90\0\x90\0\x98\0\xA0\0\xA0\0\xA0\0\xA0\0\xA0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\x80\0\xB0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xCC\0\xD1\0\xC0\0\xC4\0\xD1\0\xC8\0\xD1\0\xDD\0\x90\0\xED\0\xA0\0\xA0\0\xF8\0D\0L\0L\0L\0L\0L\0L\0L\0l\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x02\0\x8C\0\x02\0\xAC\0\xB0\0\xAD\0\xB1\0\xAE\0\xB2\0\xAF\0\xAC\0\xB0\0\xAD\0\xB1\0\xAE\0\xB2\0\xAF\0\xBA\0") }, unsafe { zerovec::ZeroVec::from_bytes_unchecked(b"\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x02\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\x01\0\0\0\x04\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x04\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\x05\0\0\0\0\0\0\0\0\0\0\0\0\x02\x02\x02\x02\x02\x02\x02\0\0\0\0\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\x03\0\0\0\0") }, icu::properties::props::HangulSyllableType(0u8)));
         }
         #[clippy::msrv = "1.71.1"]
         impl icu_provider::DataProvider<icu::properties::provider::HangulSyllableTypeV1Marker> for $provider {
